user_info_bot/requirements.txt
python-telegram-bot==20.06
requests==2.31
beautifulsoup4==4.12.2